+++ docs/conf.py
@@ -26,7 +26,7 @@ author = 'Krzysztof Polanski, Jongeun Park'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '1.3.9'
+release = '1.3.10'
 
 
 # -- General configuration ---------------------------------------------------